pyproject.toml
[tool.poetry]
name = "kitty_nurse"
version = "0.1.0"
description = ""
authors = ["Zoltan Kohalmy <zoltankohalmy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.7"
djangorestframework = "^3.12.2"
Pillow = "^8.1.2"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.1.0"
black = {version = "^20.8b1", allow-prereleases = true}
python-dotenv = "^0.15.0"
pytest-pythonpath = "^0.7.3"
pytest-sugar = "^0.9.4"
flake8 = "^3.9.0"
django-extensions = "^3.1.1"

[tool.pytest.ini_options]
addopts = "--ds kitty_nurse.settings"
python_paths = "src"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
